update main.js & move stories to stories/Image
