+++ packages/react-components/react-image/src/stories/Image/ImageBordered.stories.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Image } from '../Image';
+import { Image } from '../../Image';
 
 export const Bordered = () => (
   <div>
