+++ packages/react-components/react-image/src/stories/Image/ImageDescription.md
