+++ packages/react-components/react-image/src/stories/Image/ImageShadow.stories.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Image } from '../Image';
+import { Image } from '../../Image';
 
 export const Shadow = () => (
   <Image shadow src="https://fabricweb.azureedge.net/fabric-website/placeholders/300x300.png" />
