+++ packages/react-components/react-image/src/stories/Image/ImageShape.stories.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Image } from '../Image';
+import { Image } from '../../Image';
 
 export const Shape = () => (
   <div style={{ display: 'flex', gap: 8 }}>
